==> 01_select_data.ipynb <==
import json
import csv
import pandas as pd
import langid
from langdetect import detect
from textblob import TextBlob

data = {
    "review_id": [],
    "rating": [],
    "review_text": [],
}

# LANGDETECT

# with open("young_adult.json") as f:
#     for line in f:
#         line_object = json.loads(line)

#         try:
        
#             if detect(line_object["review_text"]) == "en":

#                 data["book_id"].append(line_object["book_id"])
#                 data["review_id"].append(line_object["review_id"])
#                 data["rating"].append(line_object["rating"])
#                 data["review_text"].append(line_object["review_text"])
        
#         except:
#             pass


# LANGID

with open("goodreads_reviews_young_adult.json") as f:
    i = 0
    for line in f:
        line_object = json.loads(line)
        i += 1

        if i%10000 == 0:
            print(f"Progress: {i}")

        try:
        
            if langid.classify(line_object["review_text"])[0] == "en":

                data["review_id"].append(line_object["review_id"])
                data["rating"].append(line_object["rating"])
                data["review_text"].append(line_object["review_text"])
        
        except:
            pass

        if i > 200000:
            break

    print("~~~~~ FINISHED ~~~~~")

df = pd.DataFrame(data)

# df = pd.read_csv('all_reviews.csv')
# df.drop(columns=["user_id", "book_id", "date_added", "date_updated", "read_at", "started_at", "n_votes", "n_comments"], inplace=True)
df.set_index("review_id", inplace=True)
df.head()

df.shape

df.sample(100)

df["rating"].value_counts()

df_1 = df[df["rating"] == 1].sample(5000)
df_2 = df[df["rating"] == 2].sample(5000)
df_4 = df[df["rating"] == 4].sample(5000)
df_5 = df[df["rating"] == 5].sample(5000)

df_filtered = pd.concat([df_1, df_2, df_4, df_5])

df_filtered.to_csv("reviews_filtered_2.csv")

print(detect("Hizli okunuyor. Merak uyandiriyor ve sizi surukluyor, zaman gecirmek"))
print(langid.classify("What a great ending to the trilogy!! The second book was better than the first,")[0])

###################### DIFFERENT SECTION

df = pd.read_csv("all_data/all_reviews.csv")
df.head()

df.info()

df2 = pd.read_csv("all_data/reviews_filtered_2.csv")
df2.head()

df2 = df2.drop(["review_id"], axis=1)

sentiment_map = {"1": 0, "2": 0, "4":1, "5":1}
df2["sentiment"] = df2["rating"].astype(str).apply(lambda x: sentiment_map[x])

df2.head(10)

df2 = df2.rename(columns={"review_text": "review"})
df2 = df2[["review", "rating", "sentiment"]]
df2.sample(15)

df2.to_csv("all_data/filtered.csv", index=False)

df3 = pd.read_csv("all_data/filtered.csv")
df3.head()

with pd.option_context("display.max_rows", 7, "display.max_colwidth", 100, "display.colheader_justify", "left"):
    display(df2)
