==> 02_preprocessing.ipynb <==
import pandas as pd
import regex as re
import nltk
from string import digits, punctuation
from nltk import word_tokenize
from nltk.corpus import stopwords, wordnet
from nltk.tokenize import word_tokenize
from nltk.stem.porter import PorterStemmer
from nltk.stem import WordNetLemmatizer
# from gensim.parsing.preprocessing import STOPWORDS
# nltk.download('wordnet')
# nltk.download('averaged_perceptron_tagger')

Preprocess text consists of:  
Lowering all letters  
Remove punctuation  
Remove numbers  
(Remove spoiler tags)  
(Remove urls)  
Tokenization  
Remove stopwords  
Stemming  
Lemmatization  

STOPWORDS = set()

with open("stopwords.txt", "r") as f:
    lines = f.readlines()

    for line in lines:
        STOPWORDS.add(line[:-1])

STOPWORDS

def preprocess(text):
    # lower all letters
    text = text.lower()
    
    # remove spoiler tag
    spoil = "** spoiler alert **"
    text = text.replace(spoil, " ")

    # remove urls
    url = re.compile(r"https?://\S+|www\.\S+")
    text = url.sub(r"", text)

    # remove digits
    rem_digits = str.maketrans(digits, " "*len(digits))
    text = text.translate(rem_digits)

    # remove punctuation
    rem_punctuation = str.maketrans(punctuation, " "*len(punctuation))
    text = text.translate(rem_punctuation)
    text = re.sub(r'[^\w\s]',' ', text)

    # word tokenization
    tokens = word_tokenize(text)

    # stemming
    # stemmer = PorterStemmer()
    # tokens = [stemmer.stem(word) for word in tokens]

    # lemmatization
    lemmatizer = WordNetLemmatizer()
    tokens = [lemmatizer.lemmatize(word, pos="v") for word in tokens]

    # remove stopwords
    tokens = [word for word in tokens if not word in STOPWORDS]

    return tokens

text = '  this review can also be found at   i picked up uglies after wondering why it was so popular  my answer mostly lies in the book s eerie similarities to the twilight zone s infamous episode   number    looks just like you    it takes place in a dystopian society where at age     everyone undergoes a complex cosmetic procedure where you are turned into a   pretty   '

tokens = word_tokenize(text)

lemmatizer = WordNetLemmatizer()
tokens = [lemmatizer.lemmatize(word, pos="v") for word in tokens]

tokens = [word for word in tokens if not word in STOPWORDS]

tokens

def preprocess2(text):
    # remove spoiler tag
    spoil = "** spoiler alert **"
    text = text.replace(spoil, " ")

    # remove urls
    url = re.compile(r"https?://\S+|www\.\S+")
    text = url.sub(r"", text)

    # tokenization
    tokens = word_tokenize(text)

    # remove digits and punctuations
    tokens = [word for word in tokens if word.isalpha()]

    # lower all letters
    tokens = [word.lower() for word in tokens]

    # remove stopwords
    # stop_words = set(stopwords.words("english"))
    tokens = [word for word in tokens if not word in STOPWORDS]

    # stemming
    # stemmer = PorterStemmer()
    # tokens = [stemmer.stem(word) for word in tokens]

    # lemmatization
    lemmatizer = WordNetLemmatizer()
    tokens = [lemmatizer.lemmatize(word, pos="v") for word in tokens]

    return tokens


df = pd.read_csv("all_data/reviews_filtered_2.csv")
df['review_text'] = df['review_text'].astype(str)
df.head()

df['text'] = df['review_text'].apply(lambda x: preprocess(x))

df.sample(15)

sentiment_map = {"1": 0, "2": 0, "4":1, "5":1}

df["sentiment"] = df["rating"].astype(str).apply(lambda x: sentiment_map[x])

df = df.rename(columns={"review_text": "review"})
df.head()

df = df[["review", "text", "rating", "sentiment"]]
df.head()

df["sentiment"].value_counts()

df.to_csv("cleaned_data/cleaned_reviews_2.csv", index=False)

################

df = pd.read_csv("all_data/ppt_filtered.csv")
df["review"] = df["review"].astype(str)
df["token"] = df["review"].apply(lambda x: preprocess(x))
df.sample(15)

df = df[["review", "token", "rating", "sentiment"]]
df.sample(15)

df.to_csv("all_data/ppt_processed.csv", index=False)
