==> 03_desc_data.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import ast

df = pd.read_csv("all_data/ppt_processed.csv")
df.head()

df["token"] = df["token"].apply(lambda x: ast.literal_eval(x))
df.head()

df["num_words"] = df["token"].apply(lambda x: len(x))

df.info()

df["sentiment"].value_counts()

df2 = df[(df["num_words"] >= 15)]
df2 = df2.drop(["num_words"], axis=1)
df2.head()

df2.info()

df2["sentiment"].value_counts()

df3 = df2[df2["sentiment"] == 0][:7000]
df4 = df2[df2["sentiment"] == 1][:7000]
df5 = pd.concat([df3, df4])
# df5 = df5.drop(["num_words"], axis=1)
df5

with pd.option_context("display.max_rows", 8, "display.max_colwidth", 100, "display.colheader_justify", "left"):
    display(df5)

df5 = df5.rename(columns={"token": "text"})
df5.head()

df5.to_csv("cleaned_data/cleaned_reviews_4.csv", index=False)

### new preprocess

df = pd.read_csv("cleaned_data/cleaned_reviews_2.csv")
df["text"] = df["text"].apply(lambda x: ast.literal_eval(x))
df["num_words"] = df["text"].apply(lambda x: len(x))
df.head()

df2 = df[(df["num_words"] >= 15)]
df2 = df2.drop(["num_words"], axis=1)
df2.head()

df2["sentiment"].value_counts()

df3 = df2[df2["sentiment"] == 0][:7000]
df4 = df2[df2["sentiment"] == 1][:7000]
df5 = pd.concat([df3, df4])
df5

df5.to_csv("cleaned_data/cleaned_reviews_4.csv", index=False)
